refactor: abstract error handling in the controllers
